Adding tests of CurveEd25519
These files need to be modified due to the addition of CurveEd25519, that can be  handled. Curve25519 was previously marked as a bad test case which is not now the case and since it was only considered the case Ed25519 in eddsa vectors, it is necessary to add Ed25519-raw, since this is the only implementation that we are going to handle for now.

https://bugs.webkit.org/show_bug.cgi?id=246145